Move validation and cleaning into BinaryTool.cs. Move baseline comparison into DetectBinaries.cs
