+++ src/SourceBuild/content/eng/tools/BinaryToolKit/BinaryTool.cs
@@ -36,19 +36,16 @@ public class BinaryTool
             isRequired: false);
 
         // Run the tooling
-        var detectedBinaries = await DetectBinaries.ExecuteAsync(targetDirectory);
+        var detectedBinaries = await DetectBinaries.ExecuteAsync(targetDirectory, outputReportDirectory, allowedBinariesFile);
 
-        var comparedBinaries = CompareBinariesAgainstBaselines
-            .Execute(
-                detectedBinaries,
-                allowedBinariesFile,
-                outputReportDirectory,
-                targetDirectory,
-                mode);
-
-        if (mode.HasFlag(Modes.Clean))
+        if (mode.HasFlag(Modes.Validate))
         {
-            RemoveBinaries.Execute(comparedBinaries, targetDirectory);
+            ValidateBinaries(detectedBinaries, outputReportDirectory);
+        }
+
+        else if (mode.HasFlag(Modes.Clean))
+        {
+            RemoveBinaries(detectedBinaries, targetDirectory);
         }
 
         Log.LogInformation("Finished all binary tasks. Took " + (DateTime.Now - startTime).TotalSeconds + " seconds.");
@@ -90,5 +87,37 @@ public class BinaryTool
             }
         }
         return fullPath;
+    }
+
+    private static void ValidateBinaries(IEnumerable<string> newBinaries, string outputReportDirectory)
+    {
+        if (newBinaries.Any())
+        {
+            string newBinariesFile = Path.Combine(outputReportDirectory, "NewBinaries.txt");
+
+            Log.LogDebug("New binaries:");
+
+            File.WriteAllLines(newBinariesFile, newBinaries);
+
+            foreach (var binary in newBinaries)
+            {
+                Log.LogDebug($"    {binary}");
+            }
+
+            Log.LogError($"ERROR: {newBinaries.Count()} new binaries. Check '{newBinariesFile}' for details.");
+        }
+    }
+
+    private static void RemoveBinaries(IEnumerable<string> binariesToRemove, string targetDirectory)
+    {
+        Log.LogInformation($"Removing binaries from '{targetDirectory}'...");
+        
+        foreach (var binary in binariesToRemove)
+        {
+            File.Delete(Path.Combine(targetDirectory, binary));
+            Log.LogDebug($"    {binary}");
+        }
+
+        Log.LogInformation($"Finished binary removal. Removed {binariesToRemove.Count()} binaries.");
     }
 }
