+++ src/SourceBuild/content/eng/tools/BinaryToolKit/DetectBinaries.cs
@@ -13,9 +13,12 @@ public static class DetectBinaries
     private const int ChunkSize = 4096;
     private static readonly Regex GitCleanRegex = new Regex(@"Would (remove|skip)( repository)? (.*)");
 
-    public static async Task<List<string>> ExecuteAsync(string targetDirectory)
+    public static async Task<List<string>> ExecuteAsync(
+        string targetDirectory,
+        string outputReportDirectory,
+        string? allowedBinariesFile)
     {
-        Log.LogInformation($"Detecting binaries in '{targetDirectory}'...");
+        Log.LogInformation($"Detecting binaries in '{targetDirectory}' not listed in '{allowedBinariesFile}'...");
 
         var matcher = new Matcher(StringComparison.Ordinal);
         matcher.AddInclude("**/*");
@@ -29,11 +32,17 @@ public static class DetectBinaries
                 return await IsBinaryAsync(file) ? file.Substring(targetDirectory.Length + 1) : null;
             });
 
-        var binaryFiles = (await Task.WhenAll(tasks)).OfType<string>().ToList();
+        var binaryFiles = (await Task.WhenAll(tasks)).OfType<string>();
+
+        var unmatchedBinaryFiles = GetUnmatchedBinaries(
+            binaryFiles,
+            allowedBinariesFile,
+            outputReportDirectory,
+            targetDirectory).ToList();
 
         Log.LogInformation($"Finished binary detection.");
 
-        return binaryFiles;
+        return unmatchedBinaryFiles;
     }
 
     private static async Task<List<string>> GetIgnoredPatternsAsync(string targetDirectory)
@@ -139,5 +148,98 @@ public static class DetectBinaries
         }
 
         return output;
+    }
+
+    private static IEnumerable<string> GetUnmatchedBinaries(
+        IEnumerable<string> searchFiles,
+        string? allowedBinariesFile,
+        string outputReportDirectory,
+        string targetDirectory)
+    {
+        HashSet<string> unmatchedFiles = new HashSet<string>(searchFiles);
+
+        var filesToPatterns = new Dictionary<string, HashSet<string>>();
+        ParseAllowedBinariesFile(allowedBinariesFile, ref filesToPatterns);
+
+        foreach (var fileToPatterns in filesToPatterns)
+        {
+            var patterns = fileToPatterns.Value;
+            HashSet<string> unusedPatterns = new HashSet<string>(patterns);
+
+            foreach (string pattern in patterns)
+            {
+                Matcher matcher = new Matcher(StringComparison.Ordinal);
+                matcher.AddInclude(pattern);
+                
+                var matches = matcher.Match(targetDirectory, searchFiles);
+                if (matches.HasMatches)
+                {
+                    unusedPatterns.Remove(pattern);
+                    unmatchedFiles.ExceptWith(matches.Files.Select(file => file.Path));
+                }
+            }
+
+            UpdateAllowedBinariesFile(fileToPatterns.Key, outputReportDirectory, unusedPatterns);
+        }
+
+        return unmatchedFiles;
+    }
+
+    private static void ParseAllowedBinariesFile(string? file, ref Dictionary<string, HashSet<string>> result)
+    {
+        if (!File.Exists(file))
+        {
+            return;
+        }
+
+        if (!result.ContainsKey(file))
+        {
+            result[file] = new HashSet<string>();
+        }
+
+        foreach (var line in File.ReadLines(file))
+        {
+            var trimmedLine = line.Trim();
+            if (string.IsNullOrWhiteSpace(trimmedLine) || trimmedLine.StartsWith("#"))
+            {
+                continue;
+            }
+
+            if (trimmedLine.StartsWith("import:"))
+            {
+                var importFile = trimmedLine.Substring("import:".Length).Trim();
+                if (!Path.IsPathFullyQualified(importFile))
+                {
+                    var currentDirectory = Path.GetDirectoryName(file)!;
+                    importFile = Path.Combine(currentDirectory, importFile);
+                }
+                if (result.ContainsKey(importFile))
+                {
+                    Log.LogWarning($"    Duplicate import {importFile}. Skipping.");
+                    continue;
+                }
+
+                ParseAllowedBinariesFile(importFile, ref result);
+            }
+            else
+            {
+                result[file].Add(trimmedLine.Split('#')[0].Trim());
+            }
+        }
+    }
+
+    private static void UpdateAllowedBinariesFile(string? file, string outputReportDirectory, HashSet<string> unusedPatterns)
+    {
+        if(File.Exists(file) && unusedPatterns.Any())
+        {
+            var lines = File.ReadAllLines(file);
+            var newLines = lines.Where(line => !unusedPatterns.Contains(line)).ToList();
+
+            string updatedFile = Path.Combine(outputReportDirectory, "Updated" + Path.GetFileName(file));
+
+            File.WriteAllLines(updatedFile, newLines);
+
+            Log.LogInformation($"    Updated allowed binaries file '{Path.GetFileName(file)}' written to '{updatedFile}'");
+        }
     }
 }
